REFACTOR: FCM 서비스 의미 단위로 나누기
