+++ src/main/java/spot/spot/domain/notification/command/service/FcmAsyncSendingUtil.java
@@ -18,7 +18,7 @@ import spot.spot.global.logging.ColorLogger;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class FcmUtil  {
+public class FcmAsyncSendingUtil {
     private final FcmTokenRepository fcmTokenRepository;
     private final FirebaseMessaging firebaseMessaging;
     // 회원 한 명과 관련된 FCM 토큰에 메시지를 보내는 기능
@@ -37,8 +37,6 @@ public class FcmUtil  {
     }
 
 
-
-
     private Message makeMessage(String token, FcmDTO fcmDTO) {
         log.info(token);
         Notification.Builder notificationBuilder =
@@ -52,7 +50,7 @@ public class FcmUtil  {
             .build();
     }
 
-    public void sendMessage(Message message) {
+    private void sendMessage(Message message) {
         try {
             firebaseMessaging.send(message);
         } catch (FirebaseMessagingException e) {
@@ -60,29 +58,6 @@ public class FcmUtil  {
             ColorLogger.red(e.getMessage());
             e.getStackTrace();
         }
-    }
-
-    public FcmDTO makeFcmDTO(String title, String body) {
-        return FcmDTO.builder()
-            .title(title)
-            .body(body)
-            .build();
-    }
-
-    public String askRequest2ClientMsg(String attenderName, String jobName){
-        return attenderName + "님이 " + jobName + "을 해결하길 원합니다!";
-    }
-
-    public String askRequest2WorkerMsg(String workerName, String jobName) {
-        return workerName + "님! " + jobName + "을 해결해 주십쇼!";
-    }
-
-    public String getStartedJobMsg(String attenderName, String jobName){
-        return attenderName + "님이 " + jobName + "을 시작합니다!";
-    }
-
-    public String requestAcceptedBody (String owner_name, String attender_name, String jobName){
-        return owner_name + "님이 " + jobName + "에 대한 " + attender_name + "님의 요청을 승낙하셨습니다!";
     }
 
 }
